requirements.txt
accelerate==1.8.1
aiofiles==23.2.1
aiohttp==3.10.5
aiosignal==1.4.0
av==15.0.0
bitsandbytes==0.45.0
decord==0.6.0
einops==0.8.0
fsspec==2025.7.0
huggingface-hub==0.34.2
jsonlines==4.0.0
llava==0.0.1.dev0
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
requests==2.32.3
tqdm==4.66.4
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
transformers==4.54.1